+++ scratch/interp_profile.py
@@ -26,7 +26,7 @@ def get_dipole(m=None, twod=False):
     z = np.array(np.linspace(-5, 5, n), dtype=dtype)
     if twod:
         y = np.array(np.linspace(-0.1, 0.1, 2), dtype=dtype)
-    crds = coordinate.wrap_crds("Rectilinear", (('z', z), ('y', y), ('x', x)))
+    crds = coordinate.wrap_crds("nonuniform_cartesian", (('z', z), ('y', y), ('x', x)))
 
     one = np.array([1.0], dtype=dtype) #pylint: disable=W0612
     three = np.array([3.0], dtype=dtype) #pylint: disable=W0612