+++ test/test_numpy.py
@@ -33,7 +33,7 @@ def main():
     y = np.array(np.linspace(-2, 2, 30), dtype=dtype)
     z = np.array(np.linspace(-5, 5, 90), dtype=dtype)
 
-    crds = coordinate.wrap_crds("Rectilinear", (('z', z), ('y', y),
+    crds = coordinate.wrap_crds("nonuniform_cartesian", (('z', z), ('y', y),
                                                 ('x', x)))
     Z, Y, X = crds.get_crds_cc(shaped=True)